Rework secretary logic
