+++ packages/backend/Cargo.toml
@@ -24,3 +24,4 @@ rand = "0.9.1"
 hex = "0.4.3"
 futures = "0.3.31"
 serde_json = "1.0.140"
+socketioxide-core = "0.17.0"
