+++ packages/backend/src/api/mod.rs
@@ -1,5 +1,4 @@
-pub mod referee;
-pub mod secretary;
+pub mod websocket;
 pub mod v1;
 
 use crate::utils::state::{auth_middleware, AppState};
