+++ packages/backend/src/api/websocket/mod.rs
@@ -1,0 +1,3 @@
+pub mod referee;
+pub mod utils;
+pub mod secretary;
